refactor(engines): extract TwoPhaseBase for flux2-klein quant family
Slice 5 of #89 (stage-axis audit findings) / PR-C closes #92.

New `engines/_two_phase_base.py::TwoPhaseBase(ImageEngine)` finalizes the
2-phase encode/generate plumbing that was previously duplicated word-for-word
across the 3 quant engines:

- encode_and_generate()       final  (was 3×49 LOC verbatim copies)
- generate_from_embeddings()  final  (was 3×49 LOC verbatim copies)
- _teardown_encoder_phase()   final  (was 3× gc block: encoder→cpu + empty_cache + gc.collect)

New `engines/helpers.py::set_execution_device(pipe)` free function
replaces the property-on-pipe-class monkey-patch (Pipeline._execution_device
override) that was previously inlined at 4 sites within Slice-5 scope:
  - flux2_klein.py x2 (inline in load_all_on_gpu + start_generation_phase)
  - flux2_klein_fp8.py (as a per-class instance method)
  - flux2_klein_fp4.py (as a per-class instance method)

Three quant engines now inherit `TwoPhaseBase`; subclasses provide only
`_load_pipeline()` plus the engine-specific load_for_encode / encode_prompt /
load_all_on_gpu / start_generation_phase scaffolding (these stay per-engine
because compile flag, NVFP4 patching, and CPU-offload semantics diverge).

Net LOC delta in engines/flux2_klein*.py: -362 (target was -130) — new
shared files add +162 → net repo delta ~-200.

Deviation: spec named the helper `_set_execution_device`. Renamed to
`set_execution_device` (no leading underscore) because pyright strict-mode
flags underscore-prefixed module-level functions imported across modules
as private-API violations. The underscore was appropriate for the pre-
refactor instance method but conflicts with Python's "module-private"
convention for a free function imported by 3 sibling modules.

Out of scope (Slice 6 / PR-D): `pulid_flux2_klein_fp4.py` still has its
own `_set_execution_device` method — PR-D's PuLIDMixin work will refactor
that file independently.

Pending verification: T7 manual smoke parity on M₂ (RTX 5070 Ti) — seeded
generation for each of the 3 engines × 3 modes (generate, batch,
batch --two-phase) must produce bit-identical pixel arrays vs staging
reference. Procedure in artifacts/plans/89-stage-axis-pr-c-plan.mdx T7.

Refs: #89 (epic), #92 (this slice). part of #89.
