+++ src/imagecli/engines/flux2_klein.py
@@ -7,21 +7,20 @@ Batch: 2-phase (encode all prompts → generate all images, ~4 GB peak in phase 
 from __future__ import annotations
 
 import logging
-from pathlib import Path
-from typing import ClassVar
 
-from imagecli.engine import EngineCapabilities, ImageEngine
+from imagecli.engine import EngineCapabilities
+from imagecli.engines._two_phase_base import TwoPhaseBase
+from imagecli.engines.helpers import set_execution_device
 
 logger = logging.getLogger(__name__)
 
 
-class Flux2KleinEngine(ImageEngine):
+class Flux2KleinEngine(TwoPhaseBase):
     name = "flux2-klein"
     description = "FLUX.2-klein-4B — best quality for 16GB VRAM (Black Forest Labs, Nov 2025)"
     model_id = "black-forest-labs/FLUX.2-klein-4B"
     vram_gb = 8.0  # FP8 + CPU offload, peak ~7.84 GB
     capabilities = EngineCapabilities(negative_prompt=False)
-    supports_two_phase: ClassVar[bool] = True
 
     def _load_pipeline(self):
         """Shared setup: load from pretrained, quantize transformer to FP8."""
@@ -105,68 +104,9 @@ class Flux2KleinEngine(ImageEngine):
         self._pipe.text_encoder.to("cuda")  # type: ignore[attr-defined]
         self._pipe.transformer.to("cuda")  # type: ignore[attr-defined]
         self._pipe.vae.to("cuda")  # type: ignore[attr-defined]
-        self._pipe._execution_device_override = __import__("torch").device("cuda")  # type: ignore[attr-defined]
-        orig_cls = type(self._pipe)
-        if not hasattr(orig_cls, "_orig_execution_device"):
-            orig_cls._orig_execution_device = orig_cls._execution_device  # type: ignore[attr-defined]
-            orig_cls._execution_device = property(  # type: ignore[attr-defined]
-                lambda pipe: (
-                    getattr(pipe, "_execution_device_override", None)
-                    or orig_cls._orig_execution_device.fget(pipe)  # type: ignore[attr-defined]
-                )
-            )
+        set_execution_device(self._pipe)
         self._optimize_pipe(self._pipe, compile=False)
         logger.info("All components on GPU (~12 GB) — no phase switching.")
-
-    def encode_and_generate(
-        self,
-        prompt: str,
-        *,
-        width: int = 1024,
-        height: int = 1024,
-        steps: int = 50,
-        guidance: float = 4.0,
-        seed: int | None = None,
-        output_path: Path,
-        callback=None,
-    ) -> Path:
-        """Encode + generate in one shot (all-on-GPU mode)."""
-        import random
-
-        import torch
-
-        if seed is None:
-            seed = random.randint(0, 2**32 - 1)
-        generator = torch.Generator("cpu").manual_seed(seed)
-
-        if torch.cuda.is_available():
-            torch.cuda.reset_peak_memory_stats()
-
-        pipe_kwargs = {
-            "prompt": prompt,
-            "width": width,
-            "height": height,
-            "num_inference_steps": steps,
-            "guidance_scale": guidance,
-            "generator": generator,
-        }
-        if callback is not None:
-            pipe_kwargs["callback_on_step_end"] = callback
-
-        assert self._pipe is not None
-        with torch.inference_mode():
-            result = self._pipe(**pipe_kwargs)  # type: ignore[operator]
-
-        image = result.images[0]
-        return self._save_image(
-            image,
-            output_path,
-            seed=seed,
-            steps=steps,
-            guidance=guidance,
-            width=width,
-            height=height,
-        )
 
     # ── 2-phase batch ──────────────────────────────────────────────────────
 
@@ -194,84 +134,15 @@ class Flux2KleinEngine(ImageEngine):
         }
 
     def start_generation_phase(self):
-        """Phase 2 setup: offload encoder, load transformer + VAE to GPU, compile."""
-        import gc
-
-        import torch
-
+        """Phase 2 setup: offload encoder, load transformer + VAE to GPU."""
         assert self._pipe is not None
         # Free encoder VRAM
-        self._pipe.text_encoder.to("cpu")  # type: ignore[attr-defined]
-        torch.cuda.empty_cache()
-        gc.collect()
-
-        import torch
+        self._teardown_encoder_phase()
 
         # Transformer (~3.9 GB FP8) + VAE (~0.17 GB) → ~4.1 GB on GPU
         self._pipe.transformer.to("cuda")  # type: ignore[attr-defined]
         self._pipe.vae.to("cuda")  # type: ignore[attr-defined]
-        # Pipeline.device / _execution_device derives from the first registered module
-        # (text_encoder, on CPU). Monkey-patch the instance method to return cuda.
-        self._pipe._execution_device_override = torch.device("cuda")  # type: ignore[attr-defined]
-        orig_cls = type(self._pipe)
-        if not hasattr(orig_cls, "_orig_execution_device"):
-            orig_cls._orig_execution_device = orig_cls._execution_device  # type: ignore[attr-defined]
-            orig_cls._execution_device = property(  # type: ignore[attr-defined]
-                lambda pipe: (
-                    getattr(pipe, "_execution_device_override", None)
-                    or orig_cls._orig_execution_device.fget(pipe)  # type: ignore[attr-defined]
-                )
-            )
+        set_execution_device(self._pipe)
         # Skip compile: torch.compile conflicts with QLinear.forward contiguity patch.
         self._optimize_pipe(self._pipe, compile=False)
         logger.info("Generation phase ready (transformer + VAE on GPU, ~4 GB).")
-
-    def generate_from_embeddings(
-        self,
-        embeddings: dict,
-        *,
-        width: int = 1024,
-        height: int = 1024,
-        steps: int = 50,
-        guidance: float = 4.0,
-        seed: int | None = None,
-        output_path: Path,
-        callback=None,
-    ) -> Path:
-        """Generate image from pre-computed prompt embeddings."""
-        import random
-
-        import torch
-
-        if seed is None:
-            seed = random.randint(0, 2**32 - 1)
-        generator = torch.Generator("cpu").manual_seed(seed)
-
-        pipe_kwargs = {
-            "prompt_embeds": embeddings["prompt_embeds"].to("cuda"),
-            "width": width,
-            "height": height,
-            "num_inference_steps": steps,
-            "guidance_scale": guidance,
-            "generator": generator,
-        }
-        if callback is not None:
-            pipe_kwargs["callback_on_step_end"] = callback
-
-        if torch.cuda.is_available():
-            torch.cuda.reset_peak_memory_stats()
-
-        assert self._pipe is not None
-        with torch.inference_mode():
-            result = self._pipe(**pipe_kwargs)  # type: ignore[operator]
-
-        image = result.images[0]
-        return self._save_image(
-            image,
-            output_path,
-            seed=seed,
-            steps=steps,
-            guidance=guidance,
-            width=width,
-            height=height,
-        )
